Implement stats service with endpoint tracking and PostgreSQL integration

+++ stats-service/stats-dto/src/main/java/ru/practicum/stats/dto/ViewStatsDto.java
@@ -1,4 +1,4 @@
 package ru.practicum.stats.dto;
 
-public record ViewStatsDto(String app, String uri, Integer hits) {
+public record ViewStatsDto(String app, String uri, Long hits) {
 }
